+++ src/api/index.ts
@@ -51,3 +51,4 @@ export { FeatureFlagsType } from './response-types/feature-flags';
 export { FeatureFlagsAPI } from './feature-flags';
 export { ContainerDistributionAPI } from './container-distribution';
 export { ExecutionEnvironmentNamespaceAPI } from './execution-environment-namespace';
+export { TaskManagementAPI } from './task-management';
