+++ src/api/response-types/task.ts
@@ -1,0 +1,10 @@
+export class TaskType {
+    pulp_created: string;
+    state: string;
+    name: string;
+    started_at: string;
+    finished_at: string;
+    error: string;
+    parent_task: string;
+    child_tasks: string;
+}
