+++ src/api/task-management.ts
@@ -1,0 +1,11 @@
+import { PulpAPI } from './pulp';
+
+export class API extends PulpAPI {
+  apiPath = 'tasks/';
+
+  constructor() {
+    super();
+  }
+}
+
+export const TaskManagementAPI = new API();
